Redefine Link_list as concatenation.
When passing a param+call list, I was intending to concatenate
two lists, not cons a single element to a list. Since cons can be
thought as a special case of concatenation. we can simply redefine
List_link as the concatenation.
Also, redefine variable in the type-specific macros in ir.h
to reflect this.

Bug reported by Allan Schottler.